Fix: Address multiple bugs and improve stability
This commit addresses several issues to improve the stability and correctness of MegaDumper:

- Fixes a 64-bit pointer arithmetic bug in MemoryAlloc.cs that could lead to IndexOutOfRangeException and other memory corruption issues. All Pointer.ToInt32() calls have been replaced with Pointer.ToInt64() for safe 64-bit address handling.

- Corrects a bug in MiniDmp.cs where the dumper was targeting the MegaDumper process itself instead of the user-selected process.

- Replaces an unconventional 	ry-catch block in GenerateDmp.cs with proper error handling to ensure exceptions during the dump process are caught and reported to the user.

- Fixes a minor typo in a message box in GenerateDmp.cs.

+++ MegaDumper/Utils/MemoryAlloc.cs
@@ -90,7 +90,7 @@ namespace Native.Memory
 
         public void ReadBytes(int offset, byte[] buffer, int startIndex, int length)
         {
-            Pointer = new IntPtr(Pointer.ToInt32() + offset);
+            Pointer = new IntPtr(Pointer.ToInt64() + offset);
             Marshal.Copy(Pointer, buffer, startIndex, length);
         }
 
@@ -126,13 +126,13 @@ namespace Native.Memory
 
         public T ReadStruct<T>(int offset, int index)
         {
-            Pointer = new IntPtr(offset + (Marshal.SizeOf(typeof(T)) * index) + Pointer.ToInt32());
+            Pointer = new IntPtr(offset + (Marshal.SizeOf(typeof(T)) * index) + Pointer.ToInt64());
             return (T)Marshal.PtrToStructure(Pointer, typeof(T));
         }
 
         public T ReadStructOffset<T>(int offset)
         {
-            Pointer = new IntPtr(Pointer.ToInt32() + offset);
+            Pointer = new IntPtr(Pointer.ToInt64() + offset);
             return (T)Marshal.PtrToStructure(Pointer, typeof(T));
         }
 
@@ -160,7 +160,7 @@ namespace Native.Memory
 
         public void WriteBytes(int offset, byte[] b)
         {
-            Pointer = new IntPtr(Pointer.ToInt32() + offset);
+            Pointer = new IntPtr(Pointer.ToInt64() + offset);
             Marshal.Copy(b, 0, Pointer, b.Length);
         }
 
@@ -191,7 +191,7 @@ namespace Native.Memory
 
         public void WriteStruct<T>(int offset, int index, T s)
         {
-            Pointer = new IntPtr(Pointer.ToInt32() + offset + (Marshal.SizeOf(typeof(T)) * index));
+            Pointer = new IntPtr(Pointer.ToInt64() + offset + (Marshal.SizeOf(typeof(T)) * index));
             Marshal.StructureToPtr(s, Pointer, false);
         }
